% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_log.R
\name{decomp_log1}
\alias{decomp_log1}
\title{Decomp and contribution calculation function for multiplicative model}
\usage{
decomp_log1(input_data, input_var, y, y_m = NULL, date, cs, sm_factor,
  is.output = F, is.index = F)
}
\arguments{
\item{input_data}{a data.table class dataset for modeling.}

\item{input_var}{a data.table class dataset for variable coefficient.}

\item{y}{a charactor for the dependent variable name in input_data.}

\item{y_m}{a charactor for the mean of dependent variable name in input_data; default is NULL. If not NULL, then is.index must be True.}

\item{date}{a charactor of the date variable name in input_data.}

\item{cs}{a charactore vector for the cross section variable names in input_data. It could be more than one variable.}

\item{sm_factor}{the smoother span. This gives the proportion of points in the plot which influence the smooth at each value. Larger values give more smoothness.}

\item{is.output}{if TRUE, then output tabels to work directory. Default is False.}

\item{is.index}{if TRUE, then y_m argument need to be setup. Default is False.}
}
\value{
a list of modeling result. It contains the following components:
   con: contribution table;
   decomp: decomp table;
   app: a shiny app;
}
\description{
A function to calucate decomp and contribution for multiplicative model
}

