% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_pass.R
\name{direct_weight}
\alias{direct_weight}
\title{direct_weight:
calculate the decomp for all true & fake models & sub-models by using the algorithm described in 'decomp_log'}
\usage{
direct_weight(input_layer = layer, input_var = var, input_data = data,
  cs = cs_var, date = date_var, is.output.final = F, model_name = NULL)
}
\value{
a list of modeling result. It contains the list of decomp for all models

mixed_approach:

a list of modeling result. It contains the list of decomp for all models
}
\description{
direct_weight:
calculate the decomp for all true & fake models & sub-models by using the algorithm described in 'decomp_log'
}

