% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decomp.R
\name{get_decomp}
\alias{get_decomp}
\title{get_decomp:}
\usage{
get_decomp(input_data, input_var, input_layer)
}
\arguments{
\item{input_data}{a data.table class dataset for modeling.}

\item{input_var}{a data.table class dataset for variable coefficient.}

\item{input_layer}{a data.table class dataset for model structure.}

\item{decomp_list}{is from the result of get_decomp, it should be a list contains all the models' decomp}

\item{structure}{is the set up file layer, which specified the structure of the model}
}
\value{
a list of modeling result. It contains the list of decomp for all models

get_reduced_model:

a list of modeling result. It contains the following components:
   contribution: contribution table;
   decomp: decomp table;
   for.plot: table for shiny;
   app: a shiny app;
}
\description{
get_decomp:
}

