% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_through.R
\name{get_log_decomp}
\alias{get_log_decomp}
\title{get_log_decomp:
calculate the decomp for all true & fake models & sub-models by using the algorithm described in 'decomp_log'}
\usage{
get_log_decomp(input_data, input_var, input_layer)
}
\arguments{
\item{input_data}{a data.table class dataset for modeling.}

\item{input_var}{a data.table class dataset for variable coefficient.}

\item{input_layer}{a data.table class dataset for model structure.}

\item{decomp_list}{is from the result of get_decomp, it should be a list contains all the models' decomp}

\item{layer}{is the set up file layer, which specified the structure of the model}

\item{input_var}{a data.table class dataset for variable coefficient.}

\item{cs}{a charactore vector for the cross section variable names in input_data. It could be more than one variable.}

\item{date}{a charactor of the date variable name in input_data.}

\item{is.output}{if TRUE, then output tabels to work directory. Default is False.}

\item{model_name}{if NULL, then output the passed-thru model on the first layer from input_layer.
Specific the model_name if need. I.E. model_name="web"}
}
\value{
a list of modeling result. It contains the list of decomp for all models

get_reduced_model:

a list of modeling result. It contains the following components:
   contribution: contribution table;
   decomp: decomp table;
   for.plot: table for shiny;
   app: a shiny app;
}
\description{
get_log_decomp:
calculate the decomp for all true & fake models & sub-models by using the algorithm described in 'decomp_log'
}

