% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_pass.R
\name{mixed_approach}
\alias{mixed_approach}
\title{Using this function when you have fake model. But lower then fake model,
you have some models to be passed thru by direct weight.}
\usage{
mixed_approach(input_data = data, input_layer = layer, input_var = var,
  is.output.final = F, model_name = NULL)
}
\description{
Using this function when you have fake model. But lower then fake model,
you have some models to be passed thru by direct weight.
}

